refactor(blocknode): split manager.go into focused files Extract storage, chart, and values concerns out of the monolithic manager.go into dedicated sibling files: - storage.go  — directory setup, PV/PVC lifecycle, path resolution - chart.go    — Helm install/upgrade/uninstall, StatefulSet and pod                 lifecycle, release version/value queries - values.go   — Helm values file computation (renderDefaultValues,                 readCustomValues), persistence override injection,                 and retention config injection manager.go now contains only the Manager struct, package-level constants, NewManager, CreateNamespace, and DeleteNamespace. Also rename the configMap local variable in injectRetentionConfig to blockNodeConfig to avoid confusion with Kubernetes ConfigMap resources.Ecd
