+++ CLAUDE.md
@@ -88,17 +88,18 @@ solo-provisioner
 │   ├── cluster install  #   Install cluster-level teleport agent (kubectl access)
 │   └── cluster uninstall#   Uninstall cluster-level teleport agent
 ├── alloy cluster        # Alloy observability cluster (install, uninstall)
-└── version              # Print version information
+├── version              # Print version information
+└── (demo)               # TUI demo command (dev/debug use only)
 ```
 
-Root persistent flags: `--config`, `--output`, `--version`/`-v`, `--log-level`, `--force`, `--skip-hardware-checks` (hidden).
+Root persistent flags: `--config`, `--output`, `--version`/`-v`, `--log-level`, `--force`, `--skip-hardware-checks` (hidden), `--verbose`/`-V` (expanded step-by-step TUI output), `--non-interactive`.
 
 Command-specific flags (on `block`, `kube`): `--profile`. Error control flags (on workflow commands like `block node`): `--stop-on-error`, `--rollback-on-error`, `--continue-on-error`.
 
 ### Business Logic (`internal/`)
 
 Key packages:
-- `internal/workflows/` — Multi-phase orchestration (preflight → setup → deploy → verify) using the [automa](https://github.com/automa-saga/automa) framework. Workflow steps live in `internal/workflows/steps/`.
+- `internal/workflows/` — Multi-phase orchestration (preflight → setup → deploy → verify) using the [automa](https://github.com/automa-saga/automa) framework. Workflow steps live in `internal/workflows/steps/`. Notification helpers live in `internal/workflows/notify/`.
 - `internal/migration/` — Scoped migration framework (startup migrations run before every CLI invocation; block-node migrations run during upgrades). See `docs/dev/migration-framework.md`.
 - `internal/state/` — Application state management (cluster state, software state, atomic writes)
 - `internal/blocknode/` — Block node provisioning logic and storage migrations
@@ -111,10 +112,12 @@ Key packages:
 - `internal/network/` — Network configuration
 - `internal/nio/` — Network I/O utilities (stdout/stderr wrappers)
 - `internal/paths/` — Path management utilities
+- `internal/proxy/` — HTTP proxy activation from configuration (sets env vars for downstream tools)
 - `internal/rsl/` — Runtime specification layer (machine, cluster, block node, teleport runtimes)
 - `internal/sysctl/` — System control parameter management
-- `internal/templates/` — Embedded template files (alloy, block-node, cilium, crio, kubeadm, metallb, sysctl, teleport)
+- `internal/templates/` — Embedded template files (alloy, block-node, cilium, crio, health, kubeadm, metallb, sysctl, teleport)
 - `internal/tomlx/` — TOML extension utilities
+- `internal/ui/` — TUI rendering: verbose/non-interactive modes, console logging suppression, prompts (`ui/prompt/`), message models, and output capture (unix/windows)
 - `internal/version/` — Version file management (VERSION, COMMIT)
 
 ### Public Packages (`pkg/`)
@@ -151,9 +154,14 @@ Uses `github.com/automa-saga/logx` (zerolog-based). Trace IDs are initialized in
 Detailed framework docs live in `docs/dev/`:
 - `migration-framework.md` — Migration system design and usage
 - `functionality-test-suite.md` — Testing approach and patterns
+- `acceptance-tests.md` — Acceptance/UAT test patterns
 - `golden-image.md` — VM golden image setup
 - `hidden-flags.md` — Undocumented/hidden flags reference
 - `effective-value-resolution.md` — Flag and config value resolution
+- `label_profiles.md` — Deployment profile label conventions
+- `proxy.md` — HTTP proxy configuration and activation
+- `tui-output.md` — TUI output modes and formatting
+- `tui-workflow-mapping.md` — Mapping of workflow steps to TUI messages
 
 ## Key Conventions
